Remove duplication of DateTimeInput and DateTimeSearch files
